requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic
pydantic-settings
python-dotenv
aioredis
itsdangerous
pywebpush
requests
beautifulsoup4
slowapi
python-multipart
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic
pydantic-settings
python-dotenv
redis
itsdangerous
slowapi
fastapi-cache2[redis]